refactor: 메서드 분리 및 리팩토링에 따른 컨트롤러 변경
