+++ src/main/java/com/theocean/fundering/domain/comment/controller/CommentController.java
@@ -2,7 +2,9 @@ package com.theocean.fundering.domain.comment.controller;
 
 import com.theocean.fundering.domain.comment.dto.CommentRequest;
 import com.theocean.fundering.domain.comment.dto.CommentResponse;
-import com.theocean.fundering.domain.comment.service.CommentService;
+import com.theocean.fundering.domain.comment.service.CreateCommentService;
+import com.theocean.fundering.domain.comment.service.DeleteCommentService;
+import com.theocean.fundering.domain.comment.service.ReadCommentService;
 import com.theocean.fundering.global.jwt.userInfo.CustomUserDetails;
 import com.theocean.fundering.global.utils.ApiUtils;
 import jakarta.validation.Valid;
@@ -17,7 +19,9 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class CommentController {
 
-    private final CommentService commentService;
+    private final CreateCommentService createCommentService;
+    private final ReadCommentService readCommentService;
+    private final DeleteCommentService deleteCommentService;
 
     // (기능) 댓글 작성
     // @PreAuthorize("hasRole('USER')") -> TODO: JWT에 role 추가 필요
@@ -28,7 +32,7 @@ public class CommentController {
             @PathVariable long postId) {
 
         Long memberId = 1L; // Long memberId = userDetails.getMember().getUserId();
-        commentService.createComment(memberId, postId, commentRequest);
+        createCommentService.createComment(memberId, postId, commentRequest);
 
         return ResponseEntity.ok(ApiUtils.success(null));
     }
@@ -43,30 +47,30 @@ public class CommentController {
             @PathVariable long commentId) {
 
         Long memberId = 1L; // Long memberId = userDetails.getMember().getUserId();
-        commentService.createSubComment(memberId, postId, commentId, commentRequest);
+        createCommentService.createSubComment(memberId, postId, commentId, commentRequest);
 
         return ResponseEntity.ok(ApiUtils.success(null));
     }
 
     // (기능) 댓글 목록 조회
     @GetMapping("/posts/{postId}/comments")
-    public ResponseEntity<?> getComments(
+    public ResponseEntity<?> readComments(
             @PathVariable long postId, @PageableDefault(size = 10) Pageable pageable) {
 
-        CommentResponse.FindAllDTO response = commentService.getComments(postId, pageable);
+        CommentResponse.FindAllDTO response = readCommentService.getComments(postId, pageable);
 
         return ResponseEntity.ok(ApiUtils.success(response));
     }
 
     // (기능) 대댓글 목록 조회
     @GetMapping("/posts/{postId}/comments/{commentId}")
-    public ResponseEntity<?> getSubComments(
+    public ResponseEntity<?> readSubComments(
             @PathVariable long postId,
             @PathVariable long commentId,
             @PageableDefault(size = 10) Pageable pageable) {
 
         CommentResponse.FindAllDTO response =
-                commentService.getSubComments(postId, commentId, pageable);
+                readCommentService.getSubComments(postId, commentId, pageable);
 
         return ResponseEntity.ok(ApiUtils.success(response));
     }
@@ -80,7 +84,7 @@ public class CommentController {
             @PathVariable long commentId) {
 
         Long memberId = 1L; // userDetails.getMember().getUserId();
-        commentService.deleteComment(memberId, postId, commentId);
+        deleteCommentService.deleteComment(memberId, postId, commentId);
 
         return ResponseEntity.ok(ApiUtils.success(null));
     }
